Relocated pairwise velocity stats functions into their own dedicated module
