+++ halotools/mock_observables/pairwise_velocities/tests/test_mean_los_velocity_vs_rp.py
@@ -5,7 +5,7 @@ from astropy.tests.helper import pytest
 from astropy.utils.misc import NumpyRNGContext
 
 from ..pairwise_velocity_stats import mean_los_velocity_vs_rp
-from .cf_helpers import generate_locus_of_3d_points
+from ...tests.cf_helpers import generate_locus_of_3d_points
 
 __all__ = ('test_mean_los_velocity_vs_rp_correctness1', 'test_mean_los_velocity_vs_rp_correctness2', 
     'test_mean_los_velocity_vs_rp_correctness3', 'test_mean_los_velocity_vs_rp_correctness4',
